// client/src/app/_models/params/Admin/contractReviewParams.ts
export class contractReviewParams {
     orderId = 0;
     orderDate = new Date();
     customerId = 0;
     customerNameLike = '';
     city ='';
     categoryId = 0;
     orderidInts: number[]=[];
     orderids: string='';
     sort = "name";
     pageNumber = 1;
     pageSize = 10;
     search: string='';
}